내가 좋아요 한글, 내가작성한글 디자인 변경
